Refactor contrast computation
